+++ packages/polaris-viz-core/loom.config.ts
@@ -1,3 +1,4 @@
+import {buildLibraryExtended} from '@shopify/loom-plugin-build-library-extended';
 import {buildLibrary} from '@shopify/loom-plugin-build-library';
 import {createPackage} from '@shopify/loom';
 
@@ -6,11 +7,13 @@ export default createPackage((pkg) => {
   pkg.entry({root: './src/index.ts'});
   pkg.use(
     buildLibrary({
+      jestTestEnvironment: 'jsdom',
       targets: 'extends @shopify/browserslist-config, node 12.22.0',
       commonjs: true,
       esmodules: true,
       esnext: true,
       rootEntrypoints: false,
     }),
+    buildLibraryExtended(),
   );
 });