+++ packages/polaris-viz-core/src/components/LineSeries/LineSeries.tsx
@@ -2,7 +2,7 @@ import React, {useMemo} from 'react';
 import type {ScaleLinear} from 'd3-scale';
 import {area as areaShape, line} from 'd3-shape';
 
-import type {Color, DataPoint, DataSeries, GradientStop} from '../../';
+import type {DataPoint, DataSeries, GradientStop} from '../../';
 import {
   useTheme,
   curveStepRounded,