+++ packages/polaris-viz-core/src/index.ts
@@ -1,5 +1,5 @@
 export * from './constants';
 export * from './utilities';
 export * from './types';
-export * from './components';
 export * from './hooks';
+export * from './components';