+++ packages/polaris-viz-core/src/utilities/index.ts
@@ -1,4 +1,5 @@
 /* eslint-disable import/no-cycle */
+export {createTheme} from './createThemes';
 export {createThemes} from './createThemes';
 export {createGradient} from './createGradient';
 export {getSeriesColors} from './getSeriesColors';