+++ packages/polaris-viz-core/src/utilities/tests/curve-step-rounded.test.ts
@@ -3,7 +3,8 @@ import {scaleLinear} from 'd3-scale';
 import {line} from 'd3-shape';
 import type {Path} from 'd3-path';
 
-import {curveStepRounded, CurveStepRounded} from '../curve-step-rounded';
+import {CurveStepRounded} from '../curveStepRounded';
+import {curveStepRounded} from '../';
 
 const dataset = [0, 20];