+++ packages/polaris-viz/src/components/index.ts
@@ -2,7 +2,6 @@ export {Point} from './Point';
 export {Crosshair} from './Crosshair';
 export {HorizontalGridLines} from './HorizontalGridLines';
 export {SparkLineChart} from './SparkLineChart';
-export type {SparkLineChartProps} from './SparkLineChart';
 export {SparkBarChart} from './SparkBarChart';
 export type {SparkBarChartProps} from './SparkBarChart';
 export {LineChart, LineChartTooltipContent} from './LineChart';
@@ -24,12 +23,9 @@ export {SkipLink} from './SkipLink';
 export {VisuallyHiddenRows} from './VisuallyHiddenRows';
 export {LinePreview} from './LinePreview';
 export {Legend} from './Legend';
-export type {PolarisVizProviderProps} from '../../../polaris-viz-core/src';
 export {TooltipWrapper} from './TooltipWrapper';
 export {SimpleBarChart} from './SimpleBarChart';
 export type {SimpleBarChartProps} from './SimpleBarChart';
 
-export {
-  PolarisVizProvider,
-  LinearGradient,
-} from '../../../polaris-viz-core/src';
+export type {PolarisVizProviderProps} from '@shopify/polaris-viz-core';
+export {PolarisVizProvider, LinearGradient} from '@shopify/polaris-viz-core';