+++ packages/polaris-viz/src/index.ts
@@ -33,6 +33,7 @@ export {
   PRINT_THEME as PolarisVizPrintTheme,
 } from './constants';
 
-export {createTheme} from './utilities/create-themes';
+export {createTheme} from '@shopify/polaris-viz-core';
+// export type {SparkLineChartProps} from '@shopify/polaris-viz-core';
 
 export type {GradientStop, Color, DataSeries, DataPoint} from './types';